TriangleMesh::slice() now accepts a vector of floats instead of doubles for consistency with mesh coordinates

+++ xs/src/TriangleMesh.hpp
@@ -30,8 +30,8 @@ class TriangleMesh
     void translate(float x, float y, float z);
     void align_to_origin();
     void rotate(double angle, Point* center);
-    void slice(const std::vector<double> &z, std::vector<Polygons>* layers);
-    void slice(const std::vector<double> &z, std::vector<ExPolygons>* layers);
+    void slice(const std::vector<float> &z, std::vector<Polygons>* layers);
+    void slice(const std::vector<float> &z, std::vector<ExPolygons>* layers);
     TriangleMeshPtrs split() const;
     void merge(const TriangleMesh* mesh);
     void horizontal_projection(ExPolygons &retval) const;
